refactor: update import paths for OAuthButtons and clean up tsconfig include
